+++ src/test/java/com/picimako/justkitting/codefolding/plugindescriptor/PluginDescriptorTagsFoldingBuilderTest.java
@@ -1,8 +1,10 @@
 //Copyright 2023 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
-package com.picimako.justkitting.codefolding;
+package com.picimako.justkitting.codefolding.plugindescriptor;
 
 import com.intellij.testFramework.TestDataPath;
+import com.picimako.justkitting.codefolding.JustKittingCodeFoldingSettings;
+import com.picimako.justkitting.codefolding.JustKittingCodeFoldingTestBase;
 
 /**
  * Integration test for {@link PluginDescriptorTagsFoldingBuilder}.
@@ -22,14 +24,19 @@ public class PluginDescriptorTagsFoldingBuilderTest extends JustKittingCodeFoldi
         doXmlTestFolding();
     }
 
-    //Folding
+    //Folding - inspections
 
     public void testPlugin() {
         JustKittingCodeFoldingSettings.getInstance().setCollapsePluginDescriptorTags(true);
         doXmlTestFolding();
     }
 
-    public void testOtherPlugin() {
+    public void testOtherLocalInspectionPlugin() {
+        JustKittingCodeFoldingSettings.getInstance().setCollapsePluginDescriptorTags(true);
+        doXmlTestFolding();
+    }
+
+    public void testOtherGlobalInspectionPlugin() {
         JustKittingCodeFoldingSettings.getInstance().setCollapsePluginDescriptorTags(true);
         doXmlTestFolding();
     }